@WIP Adding a form object for GenericWork
I want to push the edit logic into the form.

I may be able to merge this behavior with the CurationConcern::Actor
so that we can lean on the possible more obvious Form object.

Fortunately, each of the Actors and Controllers makes use of
extending a base controller, so we should be able to swap this out.